minor file location refactor
